requirements.txt
streamlit==1.51.0
pytesseract==0.3.13
pillow==10.0.0
pymupdf==1.26.6
google-generativeai==0.8.5
python-dotenv==1.0.0
